highlight folders when selected
this still doesn't work when you select one folder and then search for
something else. Ideally a folder that was selected before the search
should still show up as highlighted, but apparently returning the
highlight group in entry_maker doesn't work properly and I couldn't find
another way to do that.

It's annoying because the user could end up "selecting" a folder that's
already selected by just looking at the result panel without realizing
that it was already selected; since it was then "selecting" it will
actually remove it from the selected list, which is exactly the
opposite of what the user wanted to do